Specify custom dSYMs
Enabled drag and drop on the NSTextView. Dropped dSYM files will be the only ones used by the Symbolicator. A bar is shown at the bottom of the NSTextView so the user sees which dSYM he has specified and there he also has the ability to remove his custom dSYM file again.